+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+# [2.9.1]
+- improvements in `MoveToPositionAlongThePath`. Resolve issue [299](https://github.com/RafaelBarbosatec/bonfire/issues/299)
+- improvements in `followComponent`. Now return `true` if can move. Resolve issue [301](https://github.com/RafaelBarbosatec/bonfire/issues/301)
+
 # [2.9.0]
 - BREAKING CHANGE:
   - remove `BonfireTiledWidget`. now use `BonfireWidget` passing map `WorldMapByTiled`